+++ embassy-usb/src/class/midi.rs
@@ -1,8 +1,9 @@
 //! MIDI class implementation.
 
-use crate::Builder;
 use crate::descriptor::{SynchronizationType, UsageType};
 use crate::driver::{Driver, Endpoint, EndpointError, EndpointIn, EndpointOut, EndpointType};
+use crate::types::StringIndex;
+use crate::{Builder, Handler};
 
 /// This should be used as `device_class` when building the `UsbDevice`.
 pub const USB_AUDIO_CLASS: u8 = 0x01;
@@ -44,6 +45,18 @@ pub struct MidiClassConfig {
     /// Maximum packet size for the endpoints.
     /// For full-speed devices, the value has to be one of 8, 16, 32 or 64.
     pub max_packet_size: u16,
+
+    /// Optional labels for the IN jacks.
+    /// If set, the slice must have a length of `n_in_jacks`.
+    /// For each element with a `Some` value, a string descriptor is added to name
+    /// the jack referenced by its index.
+    pub in_jack_labels: Option<&'static [Option<MidiJackLabel>]>,
+
+    /// Optional labels for the OUT jacks.
+    /// If set, the slice must have a length of `n_out_jacks`.
+    /// For each element with a `Some` value, a string descriptor is added to name
+    /// the jack referenced by its index.
+    pub out_jack_labels: Option<&'static [Option<MidiJackLabel>]>,
 }
 
 impl Default for MidiClassConfig {
@@ -52,6 +65,55 @@ impl Default for MidiClassConfig {
             n_in_jacks: 1,
             n_out_jacks: 1,
             max_packet_size: 64,
+            in_jack_labels: None,
+            out_jack_labels: None,
+        }
+    }
+}
+
+impl Handler for MidiClassConfig {
+    fn get_string(&mut self, index: StringIndex, _lang_id: u16) -> Option<&str> {
+        if let Some(labels) = self.in_jack_labels {
+            for label in labels.iter() {
+                if let Some(label) = label
+                    && label.string_index == index
+                {
+                    return Some(label.string);
+                }
+            }
+        }
+
+        if let Some(labels) = self.out_jack_labels {
+            for label in labels.iter() {
+                if let Some(label) = label
+                    && label.string_index == index
+                {
+                    return Some(label.string);
+                }
+            }
+        }
+
+        None
+    }
+}
+
+/// Label for an IN or OUT jack.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct MidiJackLabel {
+    /// String for the label.
+    string: &'static str,
+
+    /// Index of the string descriptor associared with the label.
+    string_index: StringIndex,
+}
+
+impl MidiJackLabel {
+    /// Creates a new label with the provided string.
+    pub fn new<'d, D: Driver<'d>>(builder: &mut Builder<'d, D>, string: &'static str) -> Self {
+        Self {
+            string,
+            string_index: builder.string(),
         }
     }
 }
@@ -75,12 +137,37 @@ pub struct MidiClass<'d, D: Driver<'d>> {
 
 impl<'d, D: Driver<'d>> MidiClass<'d, D> {
     /// Creates a new `MidiClass` with the provided UsbBus and configuration.
-    pub fn new(builder: &mut Builder<'d, D>, config: MidiClassConfig) -> Self {
+    pub fn new(builder: &mut Builder<'d, D>, config: &'d mut MidiClassConfig) -> Self {
         let MidiClassConfig {
             n_in_jacks,
             n_out_jacks,
             max_packet_size,
-        } = config;
+            in_jack_labels,
+            out_jack_labels,
+        } = *config;
+
+        // Some sanity checks.
+        if n_in_jacks == 0 && n_out_jacks == 0 {
+            panic!("n_in_jacks and n_out_jacks are both 0");
+        }
+        if let Some(in_jack_labels) = in_jack_labels
+            && in_jack_labels.len() != n_in_jacks as usize
+        {
+            assert_eq!(
+                in_jack_labels.len(),
+                n_in_jacks as usize,
+                "Length of in_jack_labels slice does not match n_in_jacks"
+            );
+        }
+        if let Some(out_jack_labels) = out_jack_labels
+            && out_jack_labels.len() != n_out_jacks as usize
+        {
+            assert_eq!(
+                out_jack_labels.len(),
+                n_out_jacks as usize,
+                "Length of out_jack_labels slice does not match n_out_jacks"
+            );
+        }
 
         let mut func = builder.function(USB_AUDIO_CLASS, USB_AUDIOCONTROL_SUBCLASS, PROTOCOL_NONE);
 
@@ -93,7 +180,6 @@ impl<'d, D: Driver<'d>> MidiClass<'d, D> {
 
         // MIDIStreaming interface
         let mut iface = func.interface();
-        let _midi_if = iface.interface_number();
         let mut alt = iface.alt_setting(USB_AUDIO_CLASS, USB_MIDISTREAMING_SUBCLASS, PROTOCOL_NONE, None);
 
         let midi_streaming_total_length = 7
@@ -130,14 +216,35 @@ impl<'d, D: Driver<'d>> MidiClass<'d, D> {
         let in_jack_id_emb = |index| 2 * n_in_jacks + 2 * index + 2;
 
         for i in 0..n_in_jacks {
-            alt.descriptor(CS_INTERFACE, &[MIDI_IN_JACK_SUBTYPE, EXTERNAL, in_jack_id_ext(i), 0x00]);
+            let ijack = in_jack_labels.map_or(0, |labels| {
+                labels
+                    .get(i as usize)
+                    .map_or(0, |label| label.as_ref().map_or(0, |label| label.string_index.0))
+            });
+            alt.descriptor(
+                CS_INTERFACE,
+                &[MIDI_IN_JACK_SUBTYPE, EXTERNAL, in_jack_id_ext(i), ijack],
+            );
         }
 
         for i in 0..n_out_jacks {
-            alt.descriptor(CS_INTERFACE, &[MIDI_IN_JACK_SUBTYPE, EMBEDDED, in_jack_id_emb(i), 0x00]);
+            let ijack = out_jack_labels.map_or(0, |labels| {
+                labels
+                    .get(i as usize)
+                    .map_or(0, |label| label.as_ref().map_or(0, |label| label.string_index.0))
+            });
+            alt.descriptor(
+                CS_INTERFACE,
+                &[MIDI_IN_JACK_SUBTYPE, EMBEDDED, in_jack_id_emb(i), ijack],
+            );
         }
 
         for i in 0..n_out_jacks {
+            let ijack = out_jack_labels.map_or(0, |labels| {
+                labels
+                    .get(i as usize)
+                    .map_or(0, |label| label.as_ref().map_or(0, |label| label.string_index.0))
+            });
             alt.descriptor(
                 CS_INTERFACE,
                 &[
@@ -147,12 +254,17 @@ impl<'d, D: Driver<'d>> MidiClass<'d, D> {
                     0x01,
                     in_jack_id_emb(i),
                     0x01,
-                    0x00,
+                    ijack,
                 ],
             );
         }
 
         for i in 0..n_in_jacks {
+            let ijack = in_jack_labels.map_or(0, |labels| {
+                labels
+                    .get(i as usize)
+                    .map_or(0, |label| label.as_ref().map_or(0, |label| label.string_index.0))
+            });
             alt.descriptor(
                 CS_INTERFACE,
                 &[
@@ -162,7 +274,7 @@ impl<'d, D: Driver<'d>> MidiClass<'d, D> {
                     0x01,
                     in_jack_id_ext(i),
                     0x01,
-                    0x00,
+                    ijack,
                 ],
             );
         }
@@ -211,6 +323,10 @@ impl<'d, D: Driver<'d>> MidiClass<'d, D> {
         } else {
             None
         };
+
+        drop(func);
+
+        builder.handler(config);
 
         MidiClass { read_ep, write_ep }
     }